Feat: Implement q_size to count queue elements
This commit introduces q_size, a function that computes the total number
of elements in the queue. It iterates through the circular doubly linked
list starting from head->next and increments a counter until it loops
back to head. The iteration is performed using list_for_each, a macro
that simplifies traversing the list. This ensures an efficient and
concise way to determine the queue size.

Change-Id: Ib3e7ed1d83c622729a08068d0002e34e2c71a76d