Feat: Add shuffle based on Fisher-Yates algorithm
This commit implements a new command called 'shuffle' that randomizes
the order of queue elements using the Fisher-Yates algorithm. The
implementation begins by obtaining the queue size with q_size().

It then repeatedly selecting a random index from the unprocessed
portion of the queue. The selected node is extracted from its original
position and appended to the tail using list_add_tail. This rearrange
the nodes in-place without allocating new nodes.

Statistical analysis confirms that the final order is uniformly
distributed.

Change-Id: I8d8fdd5ae717e2518b4126295bbc0a8fd63054fd